Add a script to witness acceptance by Chrome
This change adds the witness that Chromium accepts the third-wheel mitm
proxy. It simply builds a docker image for ubuntu 18.04, installs the
certificate authority certificate, runs third-wheel and then spawns
chromium in the image to request `example.com` and verifies the content
is loaded. Since this runs in docker, no certificates are loaded into
the host machines certificate database.