Fix certificate spoofing to trick chromium
This change slightly modifies the certificate spoofing so that it is
accepted by Chromium in the `test_against_chrome.sh` script. The main
issue was that x509 version numbers are zero-based but it was written as
though it was 1-based.

Also added a debug function for printing the contents of certificates.