Quality of life improvement - macro to wrap MitmLayer in Arc
MitmLayer structs often need to be wrapped in an Arc to pass between
threads. This introduces verbose boilerplate so instead we provide a
macro to do this for the client. This requires non-mut references for
the MitmLayer trait, so client code must use interior mutability but
this was implied by the cross thread nature of the trait anyway.