Remove egregious unwraps
This commit removes the use of unwrap from certificates and proxy
handling. To do this it takes advantage of thiserror to propagate errors
from openssl, native_tls etc to clients. The only place that is left
un-fixed is http codec handling because that should be completely
replaced by a solid library like hyper.