Rewrite interface to expose full service functionality
Client code should be able to do anything with requests and responses
that it likes. This rewrite adds additional burden on client code in
order to use third-wheel, but allows more freedom in what it can do.
Clients can simply define a tower service, and a factory struct to
output one of these services given an already established connection to
the target server.